внесены правки и добавлена пара хуков
