+++ src/components/main-map/main-map.test.js
@@ -1,15 +1,15 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import {MainMap} from "./main-map";
+import MainMap from "./main-map";
 import {MOCK_OFFERS} from "../../test-data";
 
-it(`Should MainMap render correctly`, () => {
+it(`Should MainMap2 render correctly`, () => {
 
   const MainMapComponent = renderer
     .create(
         <MainMap
           offers={MOCK_OFFERS}
-          activePin={{}}
+          activePin={null}
         />)
     .toJSON();
